requirements.txt
numpy
# opencv-python-headless==4.11.0.86
pandas
matplotlib
cycler
PyYAML
scikit-learn
pillow
typing-extensions
termcolor
protobuf